le: flush CC2400's output buffer
The CC2400 occasionally retains a single byte in its output buffer after
we strobe SFSON to turn reception off. We must flush the buffer before
attempting to receive another packet. We do so by lauching a DMA without
interrupts enabled ignoring the results.

Ideally we'd be able to manually clock the data out, but I don't think
it's possible to manually toggle the data lines on the SSP.