Strip unused code from the firmware.
Add -Wl,--gc-sections to the LDFLAGS to let the linker remove the
unused functions from the .elf file. This not only reduces the size
of the .dfu in about 2K (even for the small firmwares) but also
removes that code from the .lst and .lss files.

Tested that the ubertooth_rxtx still works.