Configure DMA to transfer up to the largest possible packet size
Before this patch DMA was configured to transfer 11 x 4 = 44 bytes max.
This would not work with the largest possible adverstisment packets whose size
is:

4 + 1 + 1 + 37 + 3  = 46
AA|flg|len|data|crc

This patch increases the size of the DMA transfer and while at it replace that
constant 11 with something more descriptive.

This fixes issue #15 for me.