le: disable USB interrupts while touching USB data structures
If a USB interrupt fires while we're modifing a USB data structure (such
as enqueuing a packet in the FIFO), the interrupt handler can modify the
data structure we're touching. This is akin to a multithreaded app not
locking shared data structures. We avoid this by hackily disabling USB
interrupts for the duration that we're modifying the data structure.

For LE, this works because the host code polls Ubertooth in a tight
loop. If the poll fails, the host code silently ignores it and retries
again. This code should be rearchitected to handle USB without
interrupts in a safe period between packets.

Fixes #249