host: use average RSSI when calculating signal in LE
The previous code used determine_signal_and_noise(), which assumes that
packets are continuously streaming in (i.e., rx mode for BR). This
solution doesn't work for LE since packets are bursty. Additionally we
currently have no way of capturing noise values in LE.

We simply convert the max RSSI measured by Ubertooth and set the noise
value to INT8_MIN. It's a little gross but the best we can do.