firmware: disable individual DMA channels instead of whole controller
Previously the code shut down the entire DMA controller when it wanted
to reset state on an individual channel. This un-neighborly behavior
makes it impossible for other firmware components to make use of DMA.

This patch rejiggers things so that only individual channels are
disabled when they are not needed. The GPDMA controller fully powered on
at boot time, for a small power penalty when idle. But since when did
this project care about power savings?

Note that this is only lightly tested and possibly has weird corner
case side effects.