host: cleanly exit ubertooth-btle on ctrl-c
The previous version of this code (and the version that all other tools
use) attempts to close ubertooth and then calls exit(0) inside the
signal handler. In addition to being bad practice, it can fail if the
signal is received in the middle of handling another USB command,
leaving the Ubertooth running in an inconsistent state. The other tools
should be similarly refactored.