le: handle connection update procedure
The LE Connection Update Procedure allows devices to change their
connection parameters, mainly hop interval. This code has a major
shortcoming: if we miss the first packet after the update, the
connection is lost completely. This code and the new connection handling
code should be refactored to use the transmit window until at least one
packet is captured and an anchor is set.