le: introduce feature to NOT follow connections
This oft-requested feature uses the LE sniffing engine to log
advertising packets but NOT follow connections as they are observed.
Previously we recommended people use a hack involving filters, but the
new LE sniffing engine made that hack non-functional. Instead we now
formally support explicitly requesting not to follow connections.

Closes #327