le: lower priority of DMA interrupt
Occasionally a DMA interrupt would fire a few microseconds before the
timer1 interrupt for checking FSON status was scheduled. Due to some
weirdness in the ARM's NVIC, this was causing the timer1 interrupt
handler to never execute. This patch reduces DMA interrupt to the lowest
priority, allowing the timer1 interrupt handler to preempt the DMA
handler. See #326.

This does not completely solve the problem, as it is still possible for
the firmware to hang and become unresponsive to USB requests in
environments with a lot of BLE traffic.